+++ expipe/core.py
@@ -1,6 +1,5 @@
 import os
 import os.path as op
-import requests
 import collections
 import datetime as dt
 import quantities as pq
@@ -217,7 +216,7 @@ class Project(ExpipeObject):
         Get an entity, creating it if it doesn’t exist.
         """
         if name in self.entities:
-            return self.entities.[name]
+            return self.entities[name]
 
         return self._create_entity(name)
 
@@ -261,7 +260,7 @@ class Project(ExpipeObject):
         Get an template, creating it if it doesn’t exist.
         """
         if name in self.templates:
-            return self.templates.[name]
+            return self.templates[name]
 
         return self._create_template(name, contents)
 
@@ -712,34 +711,6 @@ class Message:
             content['datetime'] = dt.datetime.strptime(content['datetime'],
                                                        datetime_format)
         return content
-
-
-######################################################################################################
-# Backend
-######################################################################################################
-
-
-class FileSystemBackend(AbstractBackend):
-    def __init__(self, path):
-        super(FileSystemBackend, self).__init__(
-            path=path
-        )
-
-        self.path = pathlib.Path(path)
-        self.root, self.config = self.discover_config(path)
-
-    def discover_config(self, path):
-        current_path = pathlib.Path(path)
-        config_filename = current_path / "expipe.yaml"
-
-        if not os.path.exists(config_filename):
-            if current_path == pathlib.root:
-                raise Exception("ERROR: No expipe.yaml found in current folder or parents.")
-
-            return discover_config(current_path.parent())
-
-        with open(config_filename) as f:
-            return current_path, yaml.load(f)
 
 
 class Filerecord:
